almost done refactoring UI
